+++ fflas-ffpack/field/modular-balanced-float.h
@@ -55,7 +55,7 @@ namespace FFPACK {
 			modulus((Element)p),
 			half_mod( Element((p-1)/2)),
 			mhalf_mod( (Element) half_mod-modulus+1),
-			lmodulus (p)
+			lmodulus ((unsigned int)p)
 			,one(1),zero(0),mOne(-1)
 		{
 #ifdef DEBUG
